Module Bucket.Make

Parameters

module S : Bucket_intf.Serializable

Signature

type key = S.key

type value = S.value

type t

type error

val insert : t -> key -> value -> (value Core.Option.t, error) Core.Result.t

insert bucket k v binds v to k in the bucket and returns the previous binding if it was set.

val get : t -> key -> (value Core.Option.t, error) Core.Result.t

get bucket k returns the value of k if it was set.

val delete : t -> key -> (value Core.Option.t, error) Core.Result.t

delete bucket k removes k from the bucket and returns the previous binding if it was set.