Module type Bucket_intf.Bucket

type key

type value

type t

val insert : 'e. t -> key -> value -> (value, 'e) Core.Result.t

insert bucket k v binds v to k in the bucket and returns the previous binding if it was set.

val get : 'e. t -> key -> (value Core.Option.t, 'e) Core.Result.t

get bucket k returns the value of k if it was set.

val delete : 'e. t -> key -> (value Core.Option.t, 'e) Core.Result.t

delete bucket k removes k from the bucket and returns the previous binding if it was set.

val contains : 'e. t -> key -> (bool, 'e) Core.Result.t

contains bucket k returns true if the bucket contains the key.

val update_fetch : 'e. t -> key -> f:(value Core.Option.t -> value) -> (value Core.Option.t, 'e) Core.Result.t

update_fetch bucket k f checks if the binding for k exists, applies f to the binding and returns the new value.

val fetch_update : 'e. t -> key -> f:(value Core.Option.t -> value) -> (value Core.Option.t, 'e) Core.Result.t

fetch_update bucket k f checks if the binding for k exists, applies f to the binding and returns the old value.

val length : t -> int

length bucket returns the number of elements in the bucket in linear time.

val is_empty : t -> bool

is_empty bucket returns true if the bucket has no elements.

val clear : 'e. t -> (unit, 'e) Core.Result.t

clear bucket removes all the bindings in the bucket.

val max : 'e. t -> ((key * value) Core.Option.t, 'e) Core.Result.t

max bucket returns the largest key-value pair in the bucket.

val min : 'e. t -> ((key * value) Core.Option.t, 'e) Core.Result.t

min bucket returns the smallest key-value pair in the bucket.