Module Caliburn.Db

module Make (S : Db_intf.Serializable) : Db_intf.DB with type key = S.key and type value = S.value

A DB is the module used to interact with a data store that is parameterized over a module that represents a Serializable type. Given such a module, a data store can created or loaded using open and you can update the store using the various functions such as insert, delete, etc.