Module Kernel.Mealy

State machines inspired by Crem (Perone & Karachalias, 2023).

The core of this module is the type of the state machine : ('i, 'o, 's) t. This corresponds to a machine that can respond to 'i messages with 'o outputs while transitioning between states of type 's.

The module provides combinators for combining these state machines in various ways.

type ('input, 'output, 'state) t = {

initial : 'state;

action : 'state -> 'input -> 'output * 'state;

}

A mealy machine.

type ('i, 'o) s = {

action : 'i -> 'o * ('i, 'o) s;

}

A mealy machine with implicit state.

val unfold : ('i, 'o, 's) t -> ('i, 'o) s

unfold hides the state in an explicit Mealy machine.

val (>>>) : ('i, 'o, 's1) t -> ('o, 'c, 's2) t -> ('i, 'c, 's1 * 's2) t

Sequential composition. a >>> b returns a new machine that runs the machine a, takes its output and feeds it into machine b

val (***) : ('c, 'o, 's1) t -> ('c, 'd, 's2) t -> ('c, 'o * 'd, 's1 * 's2) t

Parallel composition. a *** b returns a new machine that takes a message of the same type, runs a and b in parallel and returns the output from both.