Module Caliburn.Db

module Make (S : Signatures.Common.Serializable) : sig ... end

A DB is the module used to interact with a data store that is parameterized over a module that represents a Serializable type. Given such a module, a data store can created or loaded using open and you can update the store using the various functions such as insert, delete, etc.