Module Interface.Make

A DB is the module used to interact with a data store that is parameterized over a module that represents a Serializable type. Given such a module, a data store can created or loaded using open and you can update the store using the various functions such as insert, delete, etc.

You must pass in a module that contains functions that show how to convert between the byte level representations of the types of the keys and values. A simply way to derive them is to use @@deriving bin_io

Parameters

module S : Serializable

Signature

type key = S.key

type value = S.value

type t

type errors

val sexp_of_errors : errors -> Sexplib0.Sexp.t

val errors_of_sexp : Sexplib0.Sexp.t -> errors

val open_db : string -> (t, errors) Stdlib.Result.t

open_db loads the database stored at the path into memory if it exists or creates a database at that directory and initializes a fresh instance.

val close_db : t -> unit

close_db closes the database and flushes any pending writes.